Bugfix: update results with better team formations
The function call sorted() in bestPlayer() function in generator.py is not properly applied, leading to random player selected rather than the best player.

Therefore we properly modify the list of players to ensure that the best players are being picked. This models the best effort more closely leading to better data.